backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
httpx==0.28.1
langgraph==1.1.6
pydantic==2.9.0
pydantic-core==2.23.2

